# spring-server/src/main/java/io/swagger/model/AssignStudentsRequest.java
package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import io.swagger.model.AssignStudentsRequestData;
import java.util.ArrayList;
import java.util.List;
import javax.validation.Valid;
import javax.validation.constraints.*;

/**
 * AssignStudentsRequest
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-10-10T10:52:14.903Z")

public class AssignStudentsRequest   {
  @JsonProperty("data")
  private List<AssignStudentsRequestData> data = null;

  public AssignStudentsRequest data(List<AssignStudentsRequestData> data) {
    this.data = data;
    return this;
  }

  public AssignStudentsRequest addDataItem(AssignStudentsRequestData dataItem) {
    if (this.data == null) {
      this.data = new ArrayList<AssignStudentsRequestData>();
    }
    this.data.add(dataItem);
    return this;
  }

   /**
   * Get data
   * @return data
  **/
  @ApiModelProperty(value = "")

  @Valid

  public List<AssignStudentsRequestData> getData() {
    return data;
  }

  public void setData(List<AssignStudentsRequestData> data) {
    this.data = data;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    AssignStudentsRequest assignStudentsRequest = (AssignStudentsRequest) o;
    return Objects.equals(this.data, assignStudentsRequest.data);
  }

  @Override
  public int hashCode() {
    return Objects.hash(data);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class AssignStudentsRequest {\n");
    
    sb.append("    data: ").append(toIndentedString(data)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}